frontend

+++ app/Http/Controllers/CollectionController.php
@@ -16,15 +16,13 @@ class CollectionController extends BaseController
     {
         $book = Book::where('id', $id)->get();
         $qrcodes = Qrcode::where('book_id', $id)->get();
-
         return Inertia::render('Book/Editqrcode', compact('book', 'qrcodes'));
     }
-
     public function destroy(Request $request, $id)
     {
         $qrcode = Qrcode::find($id);
         if ($qrcode) {
-            \Storage::disk('public')->delete('photos/' . basename($qrcode->photo));
+            Storage::disk('public')->delete('photos/' . basename($qrcode->photo));
             $qrcode->delete();
         }
         $book = Book::find($id);
@@ -39,12 +37,14 @@ class CollectionController extends BaseController
     }
     public function update(Request $request)
     {
+       
         $qrcode = Qrcode::find($request->id);
-        $this->service->deleteFile($qrcode->id);
-
+        if($request->hasFile('photo')){
+            $this->service->deleteFile($qrcode->id);
         $photoPath = $this->service->downloadFile($request);
         if ($photoPath) {
             $qrcode['photo']  = $photoPath;
+        }
         }
         $qrcode->fill($request->only(['publish', 'book_id', 'qr', 'year', 'ISBN', 'condition', 'booking', 'user_id']))->save();
         return Inertia::location('/collection/' . $qrcode->book_id);
